+++ src/meraxes.h
@@ -489,11 +489,11 @@ typedef struct galaxy_t
   double BlackHoleAccretedColdMass;
   double BlackHoleAccretingColdMass;
   
-  double RmetalBubble; //New for MetalEvo
+  float RmetalBubble; //New for MetalEvo
   
-  double Prefactor[70]; //here you store the prefactors of the metal bubbles
-  double Times[70]; // Time at which the SN explode!
-  double Radii[70]; 
+  float Prefactor[70]; //here you store the prefactors of the metal bubbles
+  float Times[70]; // Time at which the SN explode!
+  float Radii[70]; 
   int count_SF;
   
   // baryonic hostories
